fix #38: responds to NS/A with IP from -n argument
As described in issue #38, DNS type NS requests to iodined got responses with additional IP that was incorrect (always 2.0.0.0) if the -n argument switch was used to specify a custom IP address. The same applied to A requests (for ns.topdomain); the IP was always 2.0.0.0 because of wrong memory copying.

This commit fix solves this issue by correctly memory copying ns_ip to query destination (&q->destination).